--- src/test/java/com/academy/ui/providers/LoginFormFieldTestProvider.java
@@ -1,19 +1,0 @@
-package com.academy.ui.providers;
-
-import com.academy.utils.props.ConfigProperties;
-import org.testng.annotations.DataProvider;
-
-public class LoginFormFieldTestProvider {
-
-    ConfigProperties configProperties = new ConfigProperties();
-    String emptyFieldsErrorUA = "Потрібно заповнити всі обов'язкові поля.";
-
-    @DataProvider(name = "emptyFields")
-    public Object[][] emptyFieldsDataProvider() {
-        return new Object[][] {
-                {"Ua", configProperties.getRegisteredUserEmail(), "", emptyFieldsErrorUA},
-                {"Ua", "", configProperties.getRegisteredUserPassword(), emptyFieldsErrorUA},
-                {"Ua", "", "", emptyFieldsErrorUA},
-        };
-    }
-}
