+++ src/test/java/com/academy/ui/providers/LoginFormTestProvider.java
@@ -1,6 +1,7 @@
 package com.academy.ui.providers;
 
 import com.academy.utils.TestUtils;
+import com.academy.utils.props.ConfigProperties;
 import org.testng.annotations.DataProvider;
 
 import java.lang.reflect.Method;
@@ -12,6 +13,8 @@ public class LoginFormTestProvider {
     public LoginFormTestProvider() {
         this.testUtils = new TestUtils();
     }
+    ConfigProperties configProperties = new ConfigProperties();
+    String emptyFieldsErrorUA = "Потрібно заповнити всі обов'язкові поля.";
 
     @DataProvider(name = "verifyErrorMessageForExceedingPasswordLengthInUA")
     public Iterator<Object[]> verifyErrorMessageForExceedingPasswordLengthInUA(Method method) {
@@ -31,5 +34,14 @@ public class LoginFormTestProvider {
     @DataProvider(name = "verifyCssAndErrorIsDisplayedInForgotPasswordWithInvalidEmail")
     public Iterator<Object[]> verifyCssAndErrorIsDisplayedInForgotPasswordWithInvalidEmail(Method method) {
         return testUtils.getTestCases(method);
+    }
+
+    @DataProvider(name = "emptyFields")
+    public Object[][] emptyFieldsDataProvider() {
+        return new Object[][] {
+                {"Ua", configProperties.getRegisteredUserEmail(), "", emptyFieldsErrorUA},
+                {"Ua", "", configProperties.getRegisteredUserPassword(), emptyFieldsErrorUA},
+                {"Ua", "", "", emptyFieldsErrorUA},
+        };
     }
 }
